% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoTP.R
\name{MoTP}
\alias{MoTP}
\title{Multi-Omics Tumor Purity Prediction}
\usage{
MoTP(data_list, omics_list = c("mRNA", "miRNA", "lncRNA", "DNA-methylation"))
}
\arguments{
\item{data_list}{A list of omics data frames or matrices. Specifically, each element in the list represents a different omics dataset, where each row corresponds to a gene and each column corresponds to a tumor sample. The list must contain at least one omics dataset.}

\item{omic_list}{A character vector indicating the type of each omics data in data_list. Default is c("mRNA", "miRNA", "lncRNA", "DNA-methylation").}
}
\value{
A data frame containing sample IDs and predicted tumor purity values for each omics type as well as the averaged tumor purity across all provided omics types.
}
\description{
This function predicts tumor purity from either single-omics or multi-omics data using pre-trained models.
#'
}
\examples{
# Example usage:
# data_list <- list(mRNA_data, miRNA_data, lncRNA_data, methylation_data)
# result <- MoTP(data_list,c("mRNA", "miRNA", "lncRNA", "DNA-methylation"))
}
