% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{Preprocess}
\alias{Preprocess}
\title{Preprocess Omics Data}
\usage{
Preprocess(original_list)
}
\arguments{
\item{original_list}{A list of omics data frames or matrices in which each row is a feature and each column is a sample.The list needs to contain at least one of the omics data.}
}
\value{
The list must contain at least one omics datase
}
\description{
This function preprocessing a list of omics data frames or matrices.Including impute missing values and \link{0,1} scaled (maximum=1, minimum=0).
This step is not necessary, you can choose to process the data yourself, as long as the data has no missing values and the range is scaled to \link{0,1}.
}
\examples{
# Example usage:
# preprocessed_list <- Preprocess(original_list)
}
